Implemented packets.js copy to Mobile, backend cleanup, notes, and restructuring
The copy functionality of the Common build task utilizes shx ( https://github.com/shelljs/shx ), which allows for many useful build-script things (the 'copy' command on Windows is not crossplatform, so just using that would have made the build task Windows-only). Might want to look into making packet.js read-only to prevent mishaps with editing it ( https://github.com/shelljs/shelljs#command-reference ).

Backup work was mostly cleanup and adding more TODOs, but it is mostly stable now, so it is just implementing the packets and writing the handlers for mobile and desktop (might be able to create a shared API for both of them, might look into that sometime today or tomorrow).