Add example for secrets.config, convert database.js to class structure
database.js now has a single exported class, being 'DatabaseHandler'. This class has a constructor which will create the database from the secret.config file as we did before and store it in a field in the class. With this, all of the function in 'DatabaseHandler' don't need to be passed dbClient, as they will all work off of the field in the class.