Add Image base64 example for Node and Expo, add TypeScript node test, add UserData class and AccountType enum
Final commit of backend.js before it is migrated to TypeScript. I've tested TypeScript in Logistics-Dispatch/examples/NodeTypescriptTest/ and it seems to work fine. It will disallow 'node backend.js', but we should be executing the project through 'npx run' declared in package.json anyways, as that is how the rest of the project should be / are executed.