Converted backend.js to TypeScript, added 'start' command to package.json, added eslint rules
To execute the backend server, run 'npm run start' in the base directory. This will build the TypeScript file into a normal JavaScript file, then execute that via Node.
Converting to TypeScript allows for better visibility of any logic errors and better discrete type checking, which is kinda important when we are converting from JavaScript objects to JSON and vice-versa, as well as when we are converting from any MongoDB Documents to JSON or JavaScript objects. The extra complexity in building / executing (which is really just a proper package.json structure) is worth it for the type safety alone.