Add working packet backend logic for GET_LINKED_ORDERS with clienttest.js example, fix ObjectId in database causing errors
Packets was updated multiple times due to some bugs that didn't show until functional code was implemented, but it should work fine now. Currently, if backend is ran and clienttest is ran, a working sequence of create account -> login -> GET_LINKED_ORDERS is ran with the data from the database being returned to the client (albeit without the correct packet 'type' field, todo).
With this, mobile and desktop should be able to start calling this packet themselves and operating on the data given by the backend. More work needs to be done and the structure is kinda messy, but this can all be improved now that it is working as intended.