Desktop: Implement functioning token authentication system, implement WebSocket connection in Home.js as example
This is roughly what we are doing on mobile, where for each screen we navigate to, we will re-initialize the WebSocket connection and only handle the specific packets we need for that screen. We can re-create the WebSocket connection without any issues because we are setting the authentication token into localstorage, then using it for all the pages, so we keep our authentication with the server (potential security hazard if say a usersrcipt or cross-site exeuction script steals it from the localstorage). For instance, for an Order List screen, we wouldn't need to handle AuthenticationSuccess or failure packets, as they are not relevent to a Order List screen. However, we would want to handle SetOrderList packets and set elements of the page to that data. An example of this is in Home.js (not the appropriate place for it, so feel free to move it to it's correct home).