Modify PlaceOrder to accept preferredDate, Insert Supplier and Merchant Name into Order, Strip Password from GetUserData
PlaceOrder will now need to be passed a date string, which will then be used to determine price based on how close to that day the current day is. Currently, this is just a hardcoded value of 100 dollars, but once date is passed from desktop, this can be changed to an actual value.

Order now has the supplierName and merchantName parameters (untested), so those can now be shown instead on mobile / desktop. These should really be gotten dynamically to ensure they don't fall out-of-date in the case of a profile update, but since we cannot update profile anyways, it doesn't matter.